chore: restore global lint script
